Changed organization, started lexing
I changed the organization of the files so that the C files are located
in the src directory and header files are separated into the include
directory.

I also started working on the lexing. I started by implementing a way to
make a list of tokens, and make these tokens in one loop instead of two.
Next I will be working on the logic for the tokenize function.