Implemented removeCharacter and lexing logic
For removing characters from strings, I used a for loop, however, I
placed that for loop into a function for in place removing of
characters.

I also started working on the lexing, I am able to extract each token
string inside of a given instructional line, however, I am not able to
copy that into a c array for further processing. This commit is to just
set a ground basis, if in the future I do need to refer back to this
code.