Strange error occurring in removeCharacter function
For some reason, the removeCharacter function gets stuck on this line

mov bx, string

So it is able to parse mov and string just fine, but for bx it is not. I
located the error to be after when I remove the commas in a string.
Apparently that is too much to handle, and the comma is instead replaced
with a random character from memory. The weird thing about this is that
a couple of lines later

mov cx, [bx]

The error does not occur, and the string is parsed just fine. I persume
it might be the order that I removing characters in, or that the
function doing something weird. I will try to fix this tomorrow.