Implemented operation code parsing
The set of operations that are detailed in the readme are no implemented
into the lexer, I wrote a simple for loop to look over an array of
strings that are defined operations. For now operations and registers
can only be expressed in lowercase.

I also removed the free(section_buffer) at the end of the lexing
function, it was causing for weird characters to be placed at the end of
each section. I'm going to try and find a way to free that memory,
possibly by returning the section buffer, and freeing it in the main
function.