Parsing registers and plain values works
I've implemented the parsing of plain values and registers. I need to
work on the parsing of labels. When that is done I will work to make
this code more efficient and readable.

I've also made a function for parsing binary, as strtol was not working
with binary values too well.