Fixed issues
The address of labels will now be properly stored in a file. I need to
implement a way for the value of a label to be parsed, or have
instructions such as db dw, and so on actually put a value at their
address and have the compiler work around it when writing the file.

I also implemented a function to check if a string was only a decimal
number in util.