Register indices, label addresses
I fixed register indices, they had to be a little more complicated than
just increasing the indices uint8_t.
I also fixed the incorrect addresses of labels, by only incrementing the
address if the line is a proper instruction.

Note: There may be some problems with the startsWith function, since
it checks the length of the string that is being looked for rather then
the string that is given. I'm not sure if strncmp counters for this or
not.