Really weird issue
In test.a, when the ds is above the db, then Hello World, the value of
the DS is printed at the end of the file, in any other order, it does
not. It appears as though DS places its result into the file itself
instead of the definitions list. Thisis probably memory issues, but I am not sure
where it occurs. I am making this commit as a marker.