Weird problems
For some reason when I use version 1 or 2 of the funny pixels program
that can be found in the test folder of the emulator does not work.
The first 7 instructions are parsed properly, the rest are just NULL.
The operation is the maximum of a uint32_t and the values are 0's.
I thought it may be a memory issue, but the third version works which is
almost the same. The reason may be that I changed how compares function in
the emulator. This is very weird.